LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/background-repeat-expected.txt


FAIL Can set 'background-position' to CSS-wide keywords assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'background-position' to the 'repeat-x' keyword Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'repeat-y' keyword Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'repeat' keyword Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'space' keyword Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'round' keyword Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'no-repeat' keyword Bad value for shorthand CSS property
FAIL Setting 'background-position' to a length throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
FAIL Setting 'background-position' to a percent throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'background-position' to a time throws TypeError
PASS Setting 'background-position' to an angle throws TypeError
PASS Setting 'background-position' to a flexible length throws TypeError
FAIL Setting 'background-position' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'background-position' to a URL throws TypeError
PASS Setting 'background-position' to a transform throws TypeError
FAIL 'background-position' does not supported 'space repeat' assert_not_equals: Unsupported value must not be null got disallowed value null
